Move _remove_edge_nans function to utils module
